+++ providers/package.json
@@ -8,7 +8,8 @@
     "build": "npm run clean && npm run build:tsc",
     "build:tsc": "tsc --noEmit false",
     "clean": "rm -rf dist",
-    "test:unit": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
+    "test:unit": "jest",
+    "vitest": "vitest",
     "eject": "react-scripts eject",
     "lint": "eslint ./",
     "lint:fix": "yarn run lint -- --fix",
@@ -38,11 +39,11 @@
     "react": ">=17.0.0"
   },
   "dependencies": {
-    "@wagmi/core": "^0.7.1",
+    "@wagmi/core": "^0.6.12",
     "connectkit": "^1.0.0",
     "ethers": "^5.6.9",
     "nodemon": "^2.0.16",
-    "wagmi": "^0.8.1"
+    "wagmi": "^0.7.15"
   },
   "devDependencies": {
     "@babel/core": "^7.16.5",
